\name{dados_cvm_inf_cadastral}
\alias{dados_cvm_inf_cadastral}
\title{Package Description}
\usage{
dados_cvm_inf_cadastral(tipo)
}
\arguments{
\item{tipo}{tipo de informacao cadastral}
}
\description{
Pegas as informacoes cadastrais dos prestadores de servico do mercado financeiro disponibilizadas site http://dados.cvm.gov.br/
}
\examples{
dados_cvm_inf_cadastral("agente_autonomo_pj")
}
