\docType{data}
\name{INF_BACEN}
\alias{INF_BACEN}
\title{INF_BACEN}
\format{
Each is a tibble with two variables and three observations
}
\usage{
INF_BACEN
}
\description{
Tabela com todos os codigos do BACEN.
}
\details{
\code{INF_BACEN} asdasdaas
}
\examples{
INF_BACEN
}
\keyword{datasets}
