\name{procura_bacen}
\alias{procura_bacen}
\title{Package Description}
\usage{
procura_bacen(palavra = "")
}
\arguments{
\item{palavra}{palavra chave da pesquisa}
}
\description{
Procura o nome de todas as informacoes do BACEN disponibilizada no site https://dadosabertos.bcb.gov.br/
}
\examples{
procura_bacen(" ")
}
